# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnswerSection": {
        "additionalProperties": false,
        "properties": {
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 6,
            "title": "Claim Ids",
            "type": "array"
          },
          "kind": {
            "enum": [
              "conversation",
              "evidence",
              "general"
            ],
            "title": "Kind",
            "type": "string"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "support_label": {
            "enum": [
              "sources_checked",
              "official_lookup",
              "general_unverified",
              "conversation"
            ],
            "title": "Support Label",
            "type": "string"
          },
          "text": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "section_id",
          "kind",
          "text",
          "support_label"
        ],
        "title": "AnswerSection",
        "type": "object"
      },
...
      "ConversationChoice": {
        "additionalProperties": false,
        "properties": {
          "intent": {
            "enum": [
              "greeting",
              "identity",
              "thanks",
              "start",
              "goodbye",
              "help",
              "empathy",
              "clarification"
            ],
            "title": "Intent",
            "type": "string"
          },
          "invitation_id": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invitation Id"
          },
          "locale": {
            "enum": [
              "en",
              "zh-CN"
            ],
            "title": "Locale",
            "type": "string"
          },
          "opening_id": {
            "minimum": 0.0,
            "title": "Opening Id",
            "type": "integer"
          }
        },
        "required": [
          "locale",
          "intent",
          "opening_id"
        ],
        "title": "ConversationChoice",
        "type": "object"
      },
      "ConversationPair": {
        "additionalProperties": false,
        "properties": {
          "assistant": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Assistant",
            "type": "string"
          },
          "user": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "user",
          "assistant"
        ],
        "title": "ConversationPair",
        "type": "object"
      },
...
      "EvidencePassage": {
        "properties": {
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "classification": {
            "default": "source_statement",
            "enum": [
              "source_statement",
              "history",
              "interpretation",
              "folklore",
              "creative_adaptation"
            ],
            "title": "Classification",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "evidence_origin": {
            "default": "reviewed_corpus",
            "enum": [
              "reviewed_corpus",
              "reviewed_unit",
              "official_lookup",
              "diagnostic_candidate"
            ],
            "title": "Evidence Origin",
            "type": "string"
          },
          "exhibit_ids": {
            "items": {
              "type": "string"
            },
            "title": "Exhibit Ids",
            "type": "array"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "institution": {
            "title": "Institution",
            "type": "string"
          },
          "language": {
            "enum": [
              "en",
              "zh-CN"
            ],
            "title": "Language",
            "type": "string"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "passage_id": {
            "title": "Passage Id",
            "type": "string"
          },
          "region_ids": {
            "items": {
              "type": "string"
            },
            "title": "Region Ids",
            "type": "array"
          },
          "required_support_ids": {
            "items": {
              "type": "string"
            },
            "title": "Required Support Ids",
            "type": "array"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          },
          "rights_basis": {
            "title": "Rights Basis",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_title": {
            "title": "Source Title",
            "type": "string"
          },
          "statement_variants": {
            "items": {
              "type": "string"
            },
            "title": "Statement Variants",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "passage_id",
          "source_id",
          "text",
          "language",
          "locator",
          "rights_basis",
          "institution",
          "source_title",
          "canonical_url",
          "fetched_at",
          "reviewed_at",
          "reviewer",
          "review_id",
          "content_hash",
          "exhibit_ids",
          "region_ids"
        ],
        "title": "EvidencePassage",
        "type": "object"
      },
...
      "GuideAnswer": {
        "properties": {
          "answer_claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Answer Claim Ids",
            "type": "array"
          },
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "answer_text": {
            "title": "Answer Text",
            "type": "string"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ValidatedClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "context_claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Context Claim Ids",
            "type": "array"
          },
          "context_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Token"
          },
          "contract_version": {
            "anyOf": [
              {
                "const": "hybrid_sections_v1",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Version"
          },
          "conversation_choice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConversationChoice"
              },
              {
                "type": "null"
              }
            ]
          },
          "corpus_version": {
            "title": "Corpus Version",
            "type": "string"
          },
          "coverage_limit": {
            "title": "Coverage Limit",
            "type": "string"
          },
          "database_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Ms"
          },
          "depth": {
            "enum": [
              "concise",
              "beginner",
              "deeper"
            ],
            "title": "Depth",
            "type": "string"
          },
          "embedding_encoder_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Encoder Version"
          },
          "embedding_model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model Version"
          },
          "embedding_status": {
            "default": "disabled",
            "title": "Embedding Status",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "generation_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Ms"
          },
          "harness_version": {
            "default": "guide-hybrid-sections-v6",
            "title": "Harness Version",
            "type": "string"
          },
          "locale": {
            "enum": [
              "en",
              "zh-CN"
            ],
            "title": "Locale",
            "type": "string"
          },
          "lookup_status": {
            "default": "not_requested",
            "title": "Lookup Status",
            "type": "string"
          },
          "mode": {
            "const": "live",
            "default": "live",
            "title": "Mode",
            "type": "string"
          },
          "personality_version": {
            "default": "jinyao-conversation-v2",
            "title": "Personality Version",
            "type": "string"
          },
          "presentation_version": {
            "default": "attributed_excerpt_v1",
            "enum": [
              "attributed_excerpt_v1",
              "inventory_projection_v1",
              "conversation_composition_v1",
              "hybrid_sections_v1"
            ],
            "title": "Presentation Version",
            "type": "string"
          },
          "prompt_version": {
            "default": "jinyao-hybrid-explanations-v6",
            "title": "Prompt Version",
            "type": "string"
          },
          "provider_attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Attempt Id"
          },
          "query_embedding_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Embedding Ms"
          },
          "ranking_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Ms"
          },
          "reason": {
            "enum": [
              "reviewed_excerpt",
              "coverage_gap",
              "ambiguous_topic",
              "unsupported_claim",
              "evidence_changed",
              "unsafe_source",
              "social_turn",
              "supported_explanation",
              "general_explanation"
            ],
            "title": "Reason",
            "type": "string"
          },
          "related_exhibit_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Exhibit Ids",
            "type": "array"
          },
          "retrieval_mode": {
            "default": "lexical_only",
            "enum": [
              "lexical_only",
              "hybrid"
            ],
            "title": "Retrieval Mode",
            "type": "string"
          },
          "retrieval_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Ms"
          },
          "retrieval_version": {
            "default": "lexical-bm25-cjk-v1",
            "title": "Retrieval Version",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/AnswerSection"
            },
            "maxItems": 4,
            "title": "Sections",
            "type": "array"
          },
          "social_intent": {
            "anyOf": [
              {
                "enum": [
                  "greeting",
                  "identity",
                  "thanks",
                  "start",
                  "goodbye",
                  "help",
                  "empathy",
                  "clarification"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Social Intent"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/EvidencePassage"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "enum": [
              "answered",
              "insufficient",
              "clarification",
              "conversational"
            ],
            "title": "Status",
            "type": "string"
          },
          "uncertainty": {
            "enum": [
              "partial",
              "insufficient"
            ],
            "title": "Uncertainty",
            "type": "string"
          },
          "validation_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Ms"
          }
        },
        "required": [
          "answer_id",
          "locale",
          "depth",
          "status",
          "uncertainty",
          "reason",
          "answer_text",
          "coverage_limit",
          "corpus_version"
        ],
        "title": "GuideAnswer",
        "type": "object"
      },
      "GuideEvidenceState": {
        "properties": {
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "passage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidencePassage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "current"
        ],
        "title": "GuideEvidenceState",
        "type": "object"
      },
      "GuideRequest": {
        "additionalProperties": false,
        "properties": {
          "context_consent": {
            "default": false,
            "title": "Context Consent",
            "type": "boolean"
          },
          "context_token": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Token"
          },
          "conversation": {
            "items": {
              "$ref": "#/components/schemas/ConversationPair"
            },
            "maxItems": 6,
            "title": "Conversation",
            "type": "array"
          },
          "depth": {
            "default": "concise",
            "enum": [
              "concise",
              "beginner",
              "deeper"
            ],
            "title": "Depth",
            "type": "string"
          },
          "exhibit_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exhibit Id"
          },
          "locale": {
            "default": "en",
            "enum": [
              "en",
              "zh-CN"
            ],
            "title": "Locale",
            "type": "string"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "GuideRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidatedClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "kind": {
            "default": "source_statement",
            "enum": [
              "source_statement",
              "history",
              "interpretation",
              "folklore",
              "creative_adaptation"
            ],
            "title": "Kind",
            "type": "string"
          },
          "passage_ids": {
            "items": {
              "type": "string"
            },
            "title": "Passage Ids",
            "type": "array"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "support_method": {
            "default": "complete_reviewed_passage",
            "enum": [
              "complete_reviewed_passage",
              "reviewed_variant_v1",
              "complete_sentence_v1",
              "inventory_projection_v1",
              "official_metadata_projection_v1"
            ],
            "title": "Support Method",
            "type": "string"
          },
          "support_version": {
            "default": "deterministic-support-v1",
            "title": "Support Version",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "text",
          "passage_ids",
          "source_ids"
        ],
        "title": "ValidatedClaim",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FolkVerse API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/guide": {
      "post": {
        "operationId": "guide",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideAnswer"
                }
              },
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Anonymous session required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Request or provider limit reached"
          },
          "503": {
            "description": "Guide unavailable"
          }
        },
        "summary": "Guide"
      }
    },
    "/api/v1/guide/evidence/{identifier}": {
      "get": {
        "operationId": "inspect_guide_evidence_api_v1_guide_evidence__identifier__get",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "title": "Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideEvidenceState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Guide Evidence"
      }
    },
...
  }
}